+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.3.17
 - [plug-in] added `languages.registerCodeLensProvider` Plug-in API
+- [plug-in] added `languages.registerDocumentSymbolProvider` Plug-in API
 - [core] `ctrl+alt+a` and `ctrl+alt+d` to switch tabs left/right
 - [core] added `theia.applicationName` to application `package.json` and improved window title